requirements.txt

moviepy
edge-tts
openai-whisper
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
